Move around spec helper stuff to better fit
