+++ spec/spec_helper.rb
@@ -1,6 +1,8 @@
 ENV['ENV'] = ENV['RAILS_ENV'] = 'test'
 ENV.delete('DATABASE_URL')
 
+NOWISH = Time.now
+
 require 'travis/scheduler'
 require 'database_cleaner'
 require 'mocha'
@@ -13,6 +15,7 @@ require 'support/logger'
 require 'support/record'
 require 'support/stages'
 require 'support/rollout'
+require 'support/queues'
 
 include Mocha::API
 
@@ -64,62 +67,6 @@ RSpec.configure do |c|
     c.after(:suite) do
       puts "\nNumber of SQL queries performed:"
       puts JSON.pretty_generate(sql_count)
-    end
-  end
-
-end
-
-module SpecHelper
-  NOWISH = Time.now
-
-  def qc(*args)
-    QueueCase.new(*args)
-  end
-
-  class QueueCase
-    def initialize(cutoff: NOWISH.to_s, host: 'travis-ci.org', config: {},
-                   desc: 'uh???', queue: 'notset', education: false,
-                   created_at: NOWISH + 7.days, force_precise_sudo_required: false,
-                   force_linux_sudo_required: false)
-      @cutoff = cutoff
-      @host = host
-      @config = config
-      @desc = desc
-      @queue = queue
-      @education = education
-      @created_at = created_at
-      @force_precise_sudo_required = force_precise_sudo_required
-      @force_linux_sudo_required = force_linux_sudo_required
-    end
-
-    attr_reader :created_at, :config, :cutoff, :desc, :host, :queue
-
-    def education?
-      @education
-    end
-
-    def force_precise_sudo_required?
-      @force_precise_sudo_required
-    end
-
-    def force_linux_sudo_required?
-      @force_linux_sudo_required
-    end
-
-    def to_s
-      a = %w[when on]
-      a << (host =~ /\.org/ ? 'org' : 'com')
-      a << 'educational' if education?
-      a << "sudo=#{config[:sudo]}" if config.key?(:sudo)
-      a << "dist=#{config[:dist]}" if config.key?(:dist)
-      if force_precise_sudo_required?
-        a << "forced sudo required because of dist: precise"
-      end
-      a << "forced sudo required on linux" if force_linux_sudo_required?
-      a << 'and created'
-      a << (created_at < Time.parse(cutoff) ? "before" : "after")
-      a << 'cutoff'
-      a.join(' ')
     end
   end
 end
